Fix EmbedderLambda OOM during vector cache rebuild
Two consecutive cache-rebuild invocations were killed with
Runtime.OutOfMemory at 1024MB after ~4 minutes as the crawled corpus
grew, while the GitHub Actions job still reported "success" since
aws lambda invoke exits 0 on any response.

- scan_all() now converts each embedding to a float32 numpy array as
  soon as it's deserialized (~8x smaller than the previous float64
  Python list), the dominant memory cost at corpus scale.
- build_matrix_and_meta() builds a float32 matrix instead of float64,
  halving the cache matrix/object size.
- EmbedderLambda drops the scanned items list before writing the cache
  to reduce peak memory during the S3 write.
- Bump EmbedderLambda memory 1024MB -> 2048MB for headroom as the
  corpus continues to grow.
- run-embedder.yml now fails the job when the Lambda response carries
  an errorType, instead of reporting false "success".

+++ src/vector_store/vector_cache_store.py
@@ -39,9 +39,13 @@ def build_matrix_and_meta(items: list[dict[str, Any]]) -> tuple[np.ndarray, list
         for it in items
     ]
     if items:
-        matrix = np.asarray([it["embedding"] for it in items], dtype=np.float64)
+        # float32 halves the matrix/cache size versus float64 with no
+        # meaningful loss of cosine-similarity precision, which matters both
+        # for the EmbedderLambda rebuild's peak memory and the cache object
+        # size read back by the searcher.
+        matrix = np.asarray([it["embedding"] for it in items], dtype=np.float32)
     else:
-        matrix = np.empty((0, 0), dtype=np.float64)
+        matrix = np.empty((0, 0), dtype=np.float32)
     return matrix, meta
